requirements.txt
numpy>=1.24
pandas>=2.0
pyarrow>=12.0
PyYAML>=6.0
tqdm>=4.65
rasterio>=1.3
tifffile>=2023.7
scikit-learn>=1.3
scipy>=1.11
torch>=2.0
torchvision>=0.15
segmentation-models-pytorch>=0.3.3
